requirements.txt
aiogram==3.6.0
albumentations==1.4.6
ipython==8.12.3
matplotlib==3.7.1
numpy==1.26.4
pandas==2.0.2
Pillow==10.1.0
Pillow==10.3.0
python-dotenv==1.0.1
torch==2.1.0+cpu
torchvision==0.16.0+cpu
tqdm==4.66.1
wandb==0.17.0
